DPAD-996: adding ability to open captions list
