+++ src/experimental/players/MobileReskinnedArticleVideoPlayer/overlays/content/CaptionSelectList.tsx
@@ -1,23 +1,21 @@
 import React, { useContext } from 'react';
 import styled from 'styled-components';
-import { CaptionsSelectProps } from 'experimental/types';
+import { CaptionsSelectListProps } from 'experimental/types';
 import { PlayerContext } from 'jwplayer/players/shared/PlayerContext';
 import useCurrentCaption from 'jwplayer/utils/useCurrentCaption';
 
 const CaptionSelectContainer = styled.div`
 	position: absolute;
-	bottom: 57px;
-	right: 10px;
+	bottom: 0px;
+	right: 0px;
 	align-items: flex-start;
 	background-color: #262626;
 	flex-flow: column nowrap;
-	max-width: 284px;
 	pointer-events: auto;
-
-	height: 232px;
-	width: 284px;
-	max-height: 232px;
+	height: 100%;
+	width: 100%;
 	background: rgb(51, 51, 51);
+	z-index: 802;
 `;
 
 const CaptionsList = styled.ul``;
@@ -26,7 +24,7 @@ const CaptionTrack = styled.li<{ active: boolean }>`
 	${(props) => !props.active && `list-style-type: none;`}
 `;
 
-const CaptionsSelectList: React.FC<CaptionsSelectProps> = ({ tracksList }) => {
+const CaptionsSelectList: React.FC<CaptionsSelectListProps> = ({ tracksList }) => {
 	const { player } = useContext(PlayerContext);
 	const currentCaption = useCurrentCaption();
