+++ src/experimental/players/MobileReskinnedArticleVideoPlayer/overlays/content/CaptionsSelectOpen.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import styled from 'styled-components';
+import { CaptionsSelectOpenProps } from 'experimental/types';
+import IconBubble from '@fandom-frontend/react-common/dist/icons/IconBubble';
+
+const StyledIconBubble = styled(IconBubble)`
+	fill: #fff;
+	flex-shrink: 0;
+	height: 14px;
+	margin-right: 6px;
+	min-width: 14px;
+	width: 14px;
+`;
+
+const CaptionsSelectOpen: React.FC<CaptionsSelectOpenProps> = ({ handleOpen }) => {
+	return <StyledIconBubble onClick={handleOpen} />;
+};
+
+export default CaptionsSelectOpen;
